requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
Flask-Session==0.5.0
Werkzeug==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0
ngrok==0.12.1